Rename strings.ts to pathogenConfig.ts
