+++ src/frontend/src/common/types/pathogenConfig.ts
