+++ src/frontend/src/views/Upload/components/Review/strings.ts
@@ -1,5 +1,5 @@
 import { Pathogen } from "src/common/redux/types";
-import { PathogenConfigType } from "src/common/types/strings";
+import { PathogenConfigType } from "src/common/types/pathogenConfig";
 
 interface ReviewSamplesStrings {
   pathogenName: string;
